cleaned up, renamed etc. the German TransDM resource.
