+++ core/src/main/java/eu/excitementproject/eop/core/component/lexicalknowledge/transDm/GermanTransDmInfo.java
@@ -1,5 +1,5 @@
 
-package eu.excitementproject.eop.core.component.lexicalknowledge.dewacTransDm;
+package eu.excitementproject.eop.core.component.lexicalknowledge.transDm;
 
 import eu.excitementproject.eop.common.component.lexicalknowledge.RuleInfo;
