+++ core/src/test/java/eu/excitementproject/eop/core/GermanTransDmResourceTest.java
@@ -14,11 +14,24 @@ import eu.excitementproject.eop.common.exception.ComponentException;
 import eu.excitementproject.eop.common.exception.ConfigurationException;
 import eu.excitementproject.eop.common.representation.partofspeech.GermanPartOfSpeech;
 import eu.excitementproject.eop.common.representation.partofspeech.UnsupportedPosTagStringException;
-import eu.excitementproject.eop.core.component.lexicalknowledge.dewacTransDm.GermanTransDmInfo;
-import eu.excitementproject.eop.core.component.lexicalknowledge.dewacTransDm.GermanTransDmException;
-import eu.excitementproject.eop.core.component.lexicalknowledge.dewacTransDm.GermanTransDmResource;
+import eu.excitementproject.eop.core.component.lexicalknowledge.transDm.GermanTransDmException;
+import eu.excitementproject.eop.core.component.lexicalknowledge.transDm.GermanTransDmInfo;
+import eu.excitementproject.eop.core.component.lexicalknowledge.transDm.GermanTransDmResource;
 
-
+/**
+ * Tests the GermanTransDmResource in four different settings:
+ * 1: using all similarity measures according to a given configuration file
+ * 2: using all similarity measures by calling the resource directly
+ * 3: using only balapinc measure by calling the resource directly
+ * 4: using an invalid measure (test exception handling)
+ * 
+ * Test scenario 1 is the vastest one, by testing calls via getRulesForLeft, 
+ * getRulesForRight, and getRules, and by testing lemmas/lemma pairs which 
+ * both exist and do not exist in the resource.   
+ * 
+ * @author Britta Zeller
+ *
+ */
 public class GermanTransDmResourceTest {
 
 	
@@ -29,9 +42,9 @@ public class GermanTransDmResourceTest {
 		/** TEST 1: USE CONFIGURATION FILE */
 
 		
-		System.out.println("***************************************************************");
-		System.out.println("TEST 1");
-		System.out.println("***************************************************************");
+		//System.out.println("***************************************************************");
+		//System.out.println("TEST 1");
+		//System.out.println("***************************************************************");
 		GermanTransDmResource transDm = null;
 
 		try {
@@ -47,7 +60,7 @@ public class GermanTransDmResourceTest {
 		
 		
 		// getRulesForLeft for a N noun
-		System.out.println("lTest for Einkommen");
+		//System.out.println("lTest for Einkommen");
 		List<LexicalRule<? extends GermanTransDmInfo>> list1 = null; 
 		try {
 			list1 = transDm.getRulesForLeft("Einkommen", new GermanPartOfSpeech("N")); 
@@ -56,8 +69,8 @@ public class GermanTransDmResourceTest {
 			for (LexicalRule<? extends GermanTransDmInfo> rule : list1) {
 				assertTrue(rule.getLLemma().equals("Einkommen"));
 				if (i < 50) {
-					System.out.print("one rightLemma for 'Einkommen': " + rule.getRLemma());
-					System.out.println(" , corresponding score (" + rule.getRelation() + "): " + rule.getConfidence());
+					//System.out.print("one rightLemma for 'Einkommen': " + rule.getRLemma());
+					//System.out.println(" , corresponding score (" + rule.getRelation() + "): " + rule.getConfidence());
 				}
 				i++;
 			}						
@@ -68,16 +81,16 @@ public class GermanTransDmResourceTest {
 		}			
 
 		// NULL RESULT getRulesForLeft for a N noun
-		System.out.println("lTest for Abcdefghijk");
+		//System.out.println("lTest for Abcdefghijk");
 		List<LexicalRule<? extends GermanTransDmInfo>> list1a = null; 
 		try {
 			list1a = transDm.getRulesForLeft("Abcdefghijk", new GermanPartOfSpeech("N")); 
 			assertTrue(list1a.size() == 0);
-			System.out.println("Rule list size is " + list1a.size());
+			//System.out.println("Rule list size is " + list1a.size());
 			for (LexicalRule<? extends GermanTransDmInfo> rule : list1a) {
 				assertTrue(rule.getLLemma().equals("Abcdefghijk"));
-				System.out.print("SHOULD NOT BE SEEN: one rightLemma for 'Abcdefghijk': " + rule.getRLemma());
-				System.out.println(" , corresponding score (" + rule.getRelation() + "): " + rule.getConfidence());
+				//System.out.print("SHOULD NOT BE SEEN: one rightLemma for 'Abcdefghijk': " + rule.getRLemma());
+				//System.out.println(" , corresponding score (" + rule.getRelation() + "): " + rule.getConfidence());
 			}						
 		}
 		catch (LexicalResourceException e)
@@ -88,15 +101,15 @@ public class GermanTransDmResourceTest {
 
 		
 		// getRulesForRight for a NN noun
-		System.out.println("rTest for Truppe");
+		//System.out.println("rTest for Truppe");
 		List<LexicalRule<? extends GermanTransDmInfo>> list2 = null; 
 		try {
 			list2 = transDm.getRulesForRight("Truppe", new GermanPartOfSpeech("NN")); 
 			assertTrue(list2.size() > 0);
 			for (LexicalRule<? extends GermanTransDmInfo> rule : list2) {
 				assertTrue(rule.getRLemma().equals("Truppe"));
-				System.out.print("one leftLemma for 'Truppe': " + rule.getLLemma());
-				System.out.println(" , corresponding score (" + rule.getRelation() + "): " + rule.getConfidence());
+				//System.out.print("one leftLemma for 'Truppe': " + rule.getLLemma());
+				//System.out.println(" , corresponding score (" + rule.getRelation() + "): " + rule.getConfidence());
 			}						
 		}
 		catch (LexicalResourceException e)
@@ -107,16 +120,16 @@ public class GermanTransDmResourceTest {
 			
 
 		// NULL RESULT getRules for a J adjective + N noun
-		System.out.println("rTest for witzig");
+		//System.out.println("rTest for witzig");
 		List<LexicalRule<? extends GermanTransDmInfo>> list3 = null; 
 		try {
 			list3 = transDm.getRules("witzig", new GermanPartOfSpeech("ADJ"), "Katze", new GermanPartOfSpeech("N")); 
 			assertTrue(list3.size() == 0);
-			System.out.println("Rule list size is " + list3.size());
-			for (LexicalRule<? extends GermanTransDmInfo> rule : list3) {
-				System.out.print("SHOULD NOT BE SEEN: one leftLemma for 'witzig': " + rule.getLLemma());
-				System.out.println(" , corresponding score (" + rule.getRelation() + "): " + rule.getConfidence());
-			}						
+			//System.out.println("Rule list size is " + list3.size());
+			//for (LexicalRule<? extends GermanTransDmInfo> rule : list3) {
+				//System.out.print("SHOULD NOT BE SEEN: one leftLemma for 'witzig': " + rule.getLLemma());
+				//System.out.println(" , corresponding score (" + rule.getRelation() + "): " + rule.getConfidence());
+			//}						
 		}
 		catch (LexicalResourceException e)
 		{
@@ -125,16 +138,16 @@ public class GermanTransDmResourceTest {
 		
 		
 		// RESULT getRules for a J adjective + V verb
-		System.out.println("rTest for witzig");
+		//System.out.println("rTest for witzig");
 		List<LexicalRule<? extends GermanTransDmInfo>> list3a = null; 
 		try {
 			list3a = transDm.getRules("witzig", new GermanPartOfSpeech("ADJ"), "machen", new GermanPartOfSpeech("V")); 
 			assertTrue(list3a.size() > 0);
-			System.out.println("Rule list size is " + list3a.size());
+			//System.out.println("Rule list size is " + list3a.size());
 			for (LexicalRule<? extends GermanTransDmInfo> rule : list3a) {
 				assertTrue(rule.getLLemma().equals("witzig") ^ rule.getRLemma().equals("witzig"));
-				System.out.print("one righttLemma for 'witzig': " + rule.getRLemma());
-				System.out.println(" , corresponding score (" + rule.getRelation() + "): " + rule.getConfidence());
+				//System.out.print("one righttLemma for 'witzig': " + rule.getRLemma());
+				//System.out.println(" , corresponding score (" + rule.getRelation() + "): " + rule.getConfidence());
 			}						
 		}
 		catch (LexicalResourceException e)
@@ -167,9 +180,9 @@ public class GermanTransDmResourceTest {
 		
 		/** TEST 2: USE ALL MEASURES DIRECTLY */
 
-		System.out.println("***************************************************************");
-		System.out.println("TEST 2");
-		System.out.println("***************************************************************");
+		//System.out.println("***************************************************************");
+		//System.out.println("TEST 2");
+		//System.out.println("***************************************************************");
 		GermanTransDmResource transDmDirect1 = null;
 
 		try {
@@ -184,18 +197,18 @@ public class GermanTransDmResourceTest {
 		
 		
 		
-		System.out.println("THIS SHOULD OUTPUT THE SAME AS THE TEST BEFORE: ");
+		//System.out.println("THIS SHOULD OUTPUT THE SAME AS THE TEST BEFORE: ");
 		// RESULT getRules for a J adjective + V verb
-		System.out.println("rTest for witzig");
+		//System.out.println("rTest for witzig");
 		List<LexicalRule<? extends GermanTransDmInfo>> list5 = null; 
 		try {
 			list5 = transDmDirect1.getRules("witzig", new GermanPartOfSpeech("ADJ"), "machen", new GermanPartOfSpeech("V")); 
 			assertTrue(list5.size() > 0);
-			System.out.println("Rule list size is " + list5.size());
+			//System.out.println("Rule list size is " + list5.size());
 			for (LexicalRule<? extends GermanTransDmInfo> rule : list5) {
 				assertTrue(rule.getLLemma().equals("witzig") ^ rule.getRLemma().equals("witzig"));
-				System.out.print("one rightLemma for 'witzig': " + rule.getRLemma());
-				System.out.println(" , corresponding score (" + rule.getRelation() + "): " + rule.getConfidence());
+				//System.out.print("one rightLemma for 'witzig': " + rule.getRLemma());
+				//System.out.println(" , corresponding score (" + rule.getRelation() + "): " + rule.getConfidence());
 			}
 		}
 		catch (LexicalResourceException e)
@@ -209,9 +222,9 @@ public class GermanTransDmResourceTest {
 
 		/** TEST 3: USE ONE MEASURE DIRECTLY */
 
-		System.out.println("***************************************************************");
-		System.out.println("TEST 3");
-		System.out.println("***************************************************************");
+		//System.out.println("***************************************************************");
+		//System.out.println("TEST 3");
+		//System.out.println("***************************************************************");
 		GermanTransDmResource transDmDirect2 = null;
 
 		try {
@@ -226,18 +239,18 @@ public class GermanTransDmResourceTest {
 		
 		
 		
-		System.out.println("THIS SHOULD OUTPUT LESS RESULTS THAN THE TEST BEFORE: ");
+		//System.out.println("THIS SHOULD OUTPUT LESS RESULTS THAN THE TEST BEFORE: ");
 		// RESULT getRules for a J adjective + N noun
-		System.out.println("rTest for witzig");
+		//System.out.println("rTest for witzig");
 		List<LexicalRule<? extends GermanTransDmInfo>> list6 = null; 
 		try {
 			list6 = transDmDirect2.getRules("witzig", new GermanPartOfSpeech("ADJ"), "machen", new GermanPartOfSpeech("V")); 
 			assertTrue(list6.size() > 0);
-			System.out.println("Rule list size is " + list6.size());
+			//System.out.println("Rule list size is " + list6.size());
 			for (LexicalRule<? extends GermanTransDmInfo> rule : list6) {
 				assertTrue(rule.getLLemma().equals("witzig") ^ rule.getRLemma().equals("witzig"));
-				System.out.print("one leftLemma for 'witzig': " + rule.getLLemma());
-				System.out.println(" , corresponding score (" + rule.getRelation() + "): " + rule.getConfidence());
+				//System.out.print("one leftLemma for 'witzig': " + rule.getLLemma());
+				//System.out.println(" , corresponding score (" + rule.getRelation() + "): " + rule.getConfidence());
 			}						
 		}
 		catch (LexicalResourceException e)
@@ -250,9 +263,9 @@ public class GermanTransDmResourceTest {
 		
 		/** TEST 4: USE INVALID MEASURE DIRECTLY */
 		
-		System.out.println("***************************************************************");
-		System.out.println("TEST 4");
-		System.out.println("***************************************************************");
+		//System.out.println("***************************************************************");
+		//System.out.println("TEST 4");
+		//System.out.println("***************************************************************");
 		GermanTransDmResource transDmDirect3 = null;
 
 		try {
@@ -267,19 +280,19 @@ public class GermanTransDmResourceTest {
 		// The test halts here if transDmDirect3 is null (this should be the case)
 		Assume.assumeNotNull(transDmDirect3);
 		
-		System.out.println("THIS LINE AND FOLLOWIGN SHOULD NOT BE SEEN!");
+		//System.out.println("THIS LINE AND FOLLOWIGN SHOULD NOT BE SEEN!");
 		
 		// RESULT getRules for a J adjective + N noun
-		System.out.println("rTest for witzig");
+		//System.out.println("rTest for witzig");
 		List<LexicalRule<? extends GermanTransDmInfo>> listN = null; 
 		try {
 			listN = transDmDirect3.getRules("witzig", new GermanPartOfSpeech("ADJ"), "Einkommen", new GermanPartOfSpeech("N")); 
 			assertTrue(listN.size() > 0);
-			System.out.println("Rule list size is " + listN.size());
+			//System.out.println("Rule list size is " + listN.size());
 			for (LexicalRule<? extends GermanTransDmInfo> rule : listN) {
 				assertTrue(rule.getRLemma().equals("witzig"));
-				System.out.print("one leftLemma for 'witzig': " + rule.getLLemma());
-				System.out.println(" , corresponding score (" + rule.getRelation() + "): " + rule.getConfidence());
+				//System.out.print("one leftLemma for 'witzig': " + rule.getLLemma());
+				//System.out.println(" , corresponding score (" + rule.getRelation() + "): " + rule.getConfidence());
 			}						
 		}
 		catch (LexicalResourceException e)
